+++ src/imports/completions/filecompletions/itemfileCompletionProvider.ts
@@ -1,7 +1,6 @@
 import * as vscode from 'vscode';
 import { ItemFileObjects } from '../../../schemas/itemfileObjects';
-import { fileCompletions, getKeyObjectCompletions } from '../../utils/completionhelper';
-import { getKey, getParentKeys, isKey } from '../../utils/yamlutils';
+import { generateFileCompletion } from '../../utils/completionhelper';
 
 export function itemFileCompletionProvider(){
     return vscode.languages.registerCompletionItemProvider(
@@ -9,16 +8,8 @@ export function itemFileCompletionProvider(){
         {
             async provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, _token: vscode.CancellationToken, context: vscode.CompletionContext) {
     
-                if (context.triggerKind === vscode.CompletionTriggerKind.Invoke && isKey(document, position.line)) {
-                    const keys = getParentKeys(document, position, true).reverse();
-                    return getKeyObjectCompletions(keys.slice(1), ItemFileObjects);
-                }
+                return generateFileCompletion(document, position, context, ItemFileObjects);
 
-                if (!/^\s*$/.test(document.lineAt(position.line).text)) {
-                    return undefined;
-                }
-
-                return fileCompletions(document, position, ItemFileObjects);
             }
         }, "\n"
     );