+++ src/imports/utils/mechanicutils.ts
@@ -2,21 +2,16 @@ import { Attribute, Mechanic, ObjectInfo, ObjectType } from '../../objectInfos';
 
 
 // Utility to get mechanic data by name
-function getMechanicDataByName(name: string, type: ObjectType) {
+export function getMechanicDataByName(name: string, type: ObjectType) {
 	return ObjectInfo[type].datasetMap.get(name.toLowerCase());
 }
 
 // Utility to get all mechanics that have a name that starts with a certain string
-function getMechanicsByPrefix(prefix: string, type: ObjectType) {
+export function getMechanicsByPrefix(prefix: string, type: ObjectType) {
 	return ObjectInfo[type].dataset.filter((mechanic: Mechanic) => mechanic.name.some((name: string) => name.startsWith(prefix.toLowerCase())));
 }
 
-// Utility to get mechanic data by class
-function getMechanicDataByClass(name: string, type: ObjectType) {
-	return ObjectInfo[type].datasetClassMap.get(name.toLowerCase());
-}
-
-function getAllAttributes(mechanic: Mechanic, type: ObjectType) {
+export function getAllAttributes(mechanic: Mechanic, type: ObjectType) {
 	let attributes = mechanic.attributes;
 	if (mechanic.extends) {
 		const parentMechanic = getMechanicDataByClass(mechanic.extends, type);
@@ -37,7 +32,7 @@ function getAllAttributes(mechanic: Mechanic, type: ObjectType) {
 }
 
 // Utility to get attribute data by name
-function getAttributeDataByName(mechanic: Mechanic, attributeName: string, type: ObjectType) {
+export function getAttributeDataByName(mechanic: Mechanic, attributeName: string, type: ObjectType) {
 	const attribute = mechanic.attributes.find((attr: Attribute) => attr.name.includes(attributeName.toLowerCase()));
 	if (!attribute && mechanic.extends) {
 		const parentMechanic = getMechanicDataByClass(mechanic.extends, type);
@@ -47,6 +42,10 @@ function getAttributeDataByName(mechanic: Mechanic, attributeName: string, type:
 	}
 	return attribute;
 }
+
+
+
+
 function getInheritedAttributeDataByName(mechanic: Mechanic, attributeName: string, type: ObjectType) {
 
 	let attribute = null;
@@ -69,4 +68,7 @@ function getInheritedAttributeDataByName(mechanic: Mechanic, attributeName: stri
 	return attribute;
 }
 
-export { getMechanicDataByName, getMechanicsByPrefix, getMechanicDataByClass, getAllAttributes, getAttributeDataByName, getInheritedAttributeDataByName };
+// Utility to get mechanic data by class
+function getMechanicDataByClass(name: string, type: ObjectType) {
+	return ObjectInfo[type].datasetClassMap.get(name.toLowerCase());
+}